Move geometry functions to clkGeometry.cpp
